Eventually show newly installed apps without closing switcher
This is IMO now better than what the overview does, which is show the app name
and no icon if you search for it at the right time to mess it up.

Also switcher start up should now be fast always as we never try to refresh the
app cache while starting up. We only do it when the user hasn't typed in 500ms,
which should be a good compromise of not slowing down the normal case (in and
out of switcher as fast as possible) while still finding a time to update the
cache for new apps eventually.